app/requirements.txt
Flask==2.0.2
flask_restful==0.3.9
azure.cosmos==4.2.0